requirements.txt
absl-py
accelerate
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
astunparse
async-timeout
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
distro
exceptiongroup
filelock
flatbuffers
frozenlist
fsspec
gast
gitdb
gitpython
google
google-api-core
google-auth
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httpcore
httpx
httpx-sse
huggingface-hub
idna
jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
keras
langchain
langchain-community
langchain-core
langchain-google-firestore
langchain-openai
langchain-text-splitters
langsmith
libclang
markdown
markdown-it-py
markupsafe
marshmallow
mdurl
ml-dtypes
more-itertools
mpmath
multidict
mypy-extensions
namex
narwhals
networkx
numpy
openai
opt-einsum
optree
orjson
packaging
pandas
pillow
propcache
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydantic-settings
pydeck
pygments
python-dateutil
python-dotenv
pytz
pyyaml
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
safetensors
setuptools
six
smmap
sniffio
soupsieve
sqlalchemy
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
tiktoken
tokenizers
toml
torch
tornado
tqdm
transformers
typing-extensions
typing-inspect
typing-inspection
tzdata
urllib3
watchdog
werkzeug
wheel
wrapt
yarl
zstandard
